Changed intensity values from Double to Float, for consistency

+++ msdk-featdet/msdk-featdet-chromatogramtofeaturetable/src/main/java/io/github/msdk/featdet/chromatogramtofeaturetable/ChromatogramToFeatureTableMethod.java
@@ -204,7 +204,7 @@ public class ChromatogramToFeatureTableMethod
         if (area != null)
             row.setData(column, area);
 
-        Double height = ChromatogramUtil.getMaxHeight(intensityBuffer,
+        Float height = ChromatogramUtil.getMaxHeight(intensityBuffer,
                 numOfDataPoints);
         column = tableColumns.get(ColumnName.HEIGHT);
         if (height != null)
